feat: download bb binaries at runtime
