+++ Cargo.toml
@@ -20,6 +20,15 @@ base64 = "0.21.2"
 dirs = "5.0.1"
 tempfile = "3.6.0"
 
+## bb binary downloading
+const_format = "0.2.30"
+tar = "~0.4.15"
+flate2 = "~1.0.1"
+reqwest = { version = "0.11.16", default-features = false, features = [
+    "rustls-tls",
+    "blocking",
+] }
+
 # Native
 barretenberg-sys = { version = "0.2.0", optional = true }
 
@@ -33,17 +42,8 @@ rust-embed = { version = "6.6.0", optional = true, features = [
 wasmer = { version = "3.3", optional = true }
 
 [build-dependencies]
-dirs = "5.0.1"
 build-target = "0.4.0"
 pkg-config = "0.3"
-const_format = "0.2.30"
-tar = "~0.4.15"
-flate2 = "~1.0.1"
-reqwest = { version = "0.11.16", default-features = false, features = [
-    "rustls-tls",
-    "blocking",
-] }
-tempdir = "0.3.7"
 
 [features]
 default = ["native"]
